[#4] Rename: LoginPage -> Login
