+++ src/app/App.tsx
@@ -1,5 +1,5 @@
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import { LoginPage } from '@/pages/LoginPage';
+import { LoginPage } from '@/pages/Login';
 
 function App() {
   return (
